Large refactor for organization - Move FunctionState, metaclass, and tokenizer into their own module - Document a bit better with more to come - Prep more with the move commands ( #6 )
